requirements.txt
psl-dns
nextcord
python-dotenv
onami
aiohttp
psycopg2-binary
nextcord-ext-help-commands
nextcord-ext-menus
python-whois
forex-python
setuptools
aiohttp
aiosignal
aiosqlite
attrs
cfgv
distlib
filelock
frozenlist
identify
idna
multidict
nodeenv
platformdirs
pre-commit
pyyaml
typing-extensions
virtualenv
pymongo
motor
flask